[modify] 작성 버튼 조건부 추가 및 로그인 알림 훅 추가
