+++ src/app/layout.tsx
@@ -2,6 +2,8 @@ import type { Metadata } from "next"
 import { Inter } from "next/font/google"
 import "./globals.css"
 import Provider from "./provider"
+import FloatingButton from "@/components/ui/FloatingButton"
+import CustomModal from "@/components/CustomModal"
 
 const inter = Inter({ subsets: ["latin"] })
 
@@ -21,6 +23,8 @@ export default function RootLayout({
         <Provider>
           {children}
           <div id="portal" className="relative z-50"></div>
+          <FloatingButton />
+          <CustomModal />
         </Provider>
       </body>
     </html>
