+++ src/components/BookmarkButton.tsx
@@ -6,9 +6,8 @@ import {
 import { useMutation, useQueryClient } from "@tanstack/react-query"
 import { MdBookmarkBorder, MdOutlineBookmark } from "react-icons/md"
 import type { Tables } from "@/types/supabase"
-import { useRouter } from "next/navigation"
-import { useCustomModal } from "@/hooks/useCustomModal"
 import useUserStore from "@/store/user"
+import useLoginConfirmModal from "@/hooks/useLoginConfirmModal"
 
 interface Props {
   projectId: string
@@ -18,10 +17,7 @@ interface Props {
 const BookmarkButton = ({ projectId, bookmarks }: Props) => {
   const queryClient = useQueryClient()
   const { user: currentUser } = useUserStore((state) => state)
-
-  const router = useRouter()
-
-  const { openCustomModalHandler } = useCustomModal()
+  const openLoginConfirmModal = useLoginConfirmModal()
 
   const { mutate: addMutate } = useMutation({
     mutationFn: setBookmarks,
@@ -37,18 +33,9 @@ const BookmarkButton = ({ projectId, bookmarks }: Props) => {
   const onClickHandler = async (e: React.MouseEvent<HTMLDivElement>) => {
     e.stopPropagation()
 
-    const moveToSigninPage = () => {
-      router.push("/signin")
-    }
-
     /** 로그인 되지 않았을 때 */
     if (!currentUser) {
-      openCustomModalHandler(
-        `로그인이 필요합니다.
-        로그인 페이지로 이동하시겠습니까?`,
-        "confirm",
-        moveToSigninPage,
-      )
+      openLoginConfirmModal()
     }
 
     /** 로그인 됐을 때 */
